Move implementation from deck.h into deck.cpp
